refactor SearchResult
