+++ selenium-pages/src/main/java/com/epam/engx/selenium/pages/google/SearchResult.java
@@ -1,18 +1,16 @@
-package com.epam.engx.selenium.pages;
+package com.epam.engx.selenium.pages.google;
 
+import com.epam.engx.selenium.pages.browser.Page;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
 
 import java.time.Duration;
 import java.util.List;
 import java.util.function.Function;
 
 // page_url = https://cloud.google.com/s/results?q=Google%20Cloud%20Platform%20Pricing%20Calculator
-public class SearchResult {
-    private final WebDriver driver;
-
+public class SearchResult extends Page {
     @FindBy(css = "a > b")
     private List<WebElement> results;
 
@@ -20,9 +18,7 @@ public class SearchResult {
     private WebElement firstResultLink;
 
     public SearchResult(WebDriver driver) {
-        this.driver = driver;
-        PageFactory.initElements(driver, this);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+        super(driver);
     }
 
     public List<Link> links() {
@@ -32,5 +28,11 @@ public class SearchResult {
     public <T> T goFirst(Function<? super WebDriver, T> pageFabric) {
         firstResultLink.click();
         return pageFabric.apply(driver);
+    }
+
+    @Override
+    public SearchResult to() {
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+        return this;
     }
 }
